need to put stack after relocate :(
When objcopy creates the binary it leaves space for the stack in the
binary. Moving the stack after the relocate section puts the relocate
section immediately after the text sections in the binary, which is
where it needs to be.

I'm not sure if there is a way to change the elf or pass flags to
objcopy that would make this not the case.